+++ fixtures/test-data/newData.json
@@ -3,7 +3,8 @@
     "pageUrl" : "https://www.saucedemo.com/",
     "pageTitle" : "Swag Labs",
     "usernameValue" : "standard_user",
-    "userPasswordValue" : "abcd"
+    "userPasswordValue" : "abcd",
+    "errorMessage": "Username and password"
     },
 
    "validData": {
