+++ fixtures/test-data/routes.json
@@ -1,0 +1,5 @@
+{
+    "saucedemoPageUrl" : "https://www.saucedemo.com/",
+    "rahulshettyPageUrl": "https://rahulshettyacademy.com/AutomationPractice/"
+
+}
