+++ tests/rahulshetty/dropdown.spec.js
@@ -1,6 +1,6 @@
 const {test, expect} = require('@playwright/test')
-const {POManager} = require('../pageObjects/POManager')
-const dataset = require('../fixtures/test-data/dataset.json')
+const {POManager} = require('../../pageObjects/POManager')
+const dataset = require('../../fixtures/test-data/dataset.json')
 
 test(`Dropdown Test`, async({page})=>{
