+++ tests/rahulshetty/suggestionBox.spec.js
@@ -1,7 +1,7 @@
 
 const {test, expect} = require('@playwright/test')
-const {POManager} = require('../pageObjects/POManager')
-const dataset = require('../fixtures/test-data/practiceData.json')
+const {POManager} = require('../../pageObjects/POManager')
+const dataset = require('../../fixtures/test-data/practiceData.json')
 
 test(`Suggestion Box Test`, async({page})=>{
